refactor(email_service/send_email): split create_email_message and attach_files_to_message into email_util.py (SCRUM-73)
